requirements.txt
aiohttp
asyncpg
SQLAlchemy>=2.0
python-dotenv

